- Fixed 64-bit build.

+++ llc_std_string.h
@@ -8,10 +8,10 @@
 
 namespace llc
 {
-	ndstain	::llc::vcc		tovcc   (const std::string & srcstr)							{ return srcstr.size() ? ::llc::vcc{&srcstr[0], srcstr.size()} : ::llc::vcs{}; }
+	ndstain	::llc::vcc		tovcc   (const std::string & srcstr)							{ return srcstr.size() ? ::llc::vcc{&srcstr[0], (uint32_t)srcstr.size()} : ::llc::vcs{}; }
 	stainli	::llc::error_t	tovcc   (::llc::vcc		& output, const std::string & srcstr)	{ return (output = ::llc::tovcc(srcstr)).size(); }
 	stainli	::llc::error_t	tolabel (::llc::vcc		& output, const std::string & srcstr)	{ return (output = ::llc::label(::llc::tovcc(srcstr))).size(); }
-	stainli	::llc::error_t	toachar (::llc::achar	& output, const std::string & srcstr)	{ llc_necs(output.reserve(srcstr.length())); return (output = ::llc::tovcc(srcstr)).size(); }
+	stainli	::llc::error_t	toachar (::llc::achar	& output, const std::string & srcstr)	{ llc_necs(output.reserve((uint32_t)srcstr.length())); return (output = ::llc::tovcc(srcstr)).size(); }
 	stainli	::llc::error_t	append 	(::llc::achar	& output, const std::string & srcstr)	{ return output.append(::llc::tovcc(srcstr)); }
 
 	ndstain	::llc::vcc		tolabel (const std::string & srcstr) 		{ return ::llc::label(::llc::tovcc(srcstr)); }
